Now the app sends an e-mail with the notification if there is internet connection
If there is internet connection and an e-mail configured, the app sends an e-mail to yourself when the notification pops.
There is one more properties configuration commented in case the first stops working.
I had some problems with the anti-virus blocking up the code execution, the solution was to turn it off(avast).
Now the only thing remaining to do is some light customization